requirements.txt
# Encyclopedia Britannica Knowledge Graph Pipeline
# Install: pip install -r requirements.txt

# Embeddings
voyageai>=0.2.0

# Neo4j
neo4j>=5.0.0

# Utilities
tqdm>=4.0.0
python-dotenv>=1.0.0
